Add support for patching released bundles

+++ nightly/entrypoint.sh
@@ -1,9 +1,28 @@
 #!/bin/bash
 
+copyextras() {
+	if [ -d "${LIFERAY_HOME}/drivers" ]; then
+		rsync -av "${LIFERAY_HOME}/drivers/" "${LIFERAY_HOME}/tomcat/lib/ext/"
+	fi
+
+	if [ -d "${LIFERAY_HOME}/patches" ]; then
+		rsync -av "${LIFERAY_HOME}/patches/" "${LIFERAY_HOME}/patching-tool/patches/"
+
+		cd "${LIFERAY_HOME}/patching-tool"
+		./patching-tool.sh auto-discovery ..
+		./patching-tool.sh install
+		cd -
+	fi
+}
 
 downloadbranch() {
 	SHORT_NAME=$(echo $BASE_BRANCH | sed 's/ee-//g' | sed 's/\.//g')
 	NEW_BASELINE=
+
+	if [ "" != "$RELEASE_BUILD" ] || [ -d "${LIFERAY_HOME}/patches" ]; then
+		downloadreleasebuild
+		return 0
+	fi
 
 	echo "Trying build server"
 	downloadbranchbuild
@@ -81,7 +100,7 @@ downloadbranchmirror() {
 
 	local BUILD_CANDIDATE=$(curl -s --connect-timeout 2 $REQUEST_URL | grep -o '<a href="[^"]*tomcat-7.0-[^"]*">' | cut -d'"' -f 2 | sort | tail -1)
 
-	if [ "" == "$BUILD_TIMESTAMP" ]; then
+	if [ "" == "$BUILD_CANDIDATE" ]; then
 		return 0
 	fi
 
@@ -116,6 +135,41 @@ downloadbuild() {
 	fi
 }
 
+downloadreleasebuild() {
+	local REQUEST_URL=
+
+	if [ "" == "$RELEASE_ID" ]; then
+		if [ -d $LIFERAY_HOME/patches ]; then
+			RELEASE_ID=7.0.10
+		else
+			RELEASE_ID=7.0.0-ga1
+		fi
+	fi
+
+	if [[ "$RELEASE_ID" == 7.0.10 ]]; then
+		REQUEST_URL="$LIFERAY_FILES_MIRROR/private/ee/portal/${RELEASE_ID}/"
+	else
+		REQUEST_URL="$LIFERAY_RELEASES_MIRROR/portal/${RELEASE_ID}/"
+	fi
+
+	echo "Attempting to download from ${REQUEST_URL}"
+
+	local BUILD_CANDIDATE=$(curl -s --connect-timeout 2 $REQUEST_URL | grep -o '<a href="[^"]*tomcat-7.0-[^"]*">' | cut -d'"' -f 2 | sort | tail -1)
+
+	if [ "" == "$BUILD_CANDIDATE" ]; then
+		return 0
+	fi
+
+	REQUEST_URL="${REQUEST_URL}${BUILD_CANDIDATE}"
+
+	BUILD_TIMESTAMP=$(echo $BUILD_CANDIDATE | grep -o "[0-9]*.zip" | cut -d'.' -f 1)
+
+	echo "Downloading $RELEASE_ID release (used for patching)"
+
+	BUILD_NAME=$SHORT_NAME-$BUILD_TIMESTAMP.zip
+	getbuild $REQUEST_URL $BUILD_NAME
+}
+
 downloadtag() {
 	NEW_BASELINE=$BASE_TAG
 	BUILD_NAME=${BASE_TAG}.tar.gz
@@ -130,27 +184,31 @@ extract() {
 
 	cd ${LIFERAY_HOME}
 
+	echo "Build name: $BUILD_NAME"
+
 	if [[ "$BUILD_NAME" == *.tar.gz ]]; then
 		tar -zxf ${BUILD_NAME}
 	elif [[ "$BUILD_NAME" == *.zip ]]; then
 		unzip -qq "${BUILD_NAME}"
 	fi
 
-	local TOMCAT_NAME=$(find . -mindepth 1 -maxdepth 1 -type d -name 'tomcat*')
+	local OLD_CATALINA_HOME=$(find . -type d -name 'tomcat*')
 
-	if [ "" == "$TOMCAT_NAME" ]; then
-		local UNZIP_NAME=$(find . -mindepth 1 -maxdepth 1 -type d)
+	if [ "" != "$OLD_CATALINA_HOME" ]; then
+		local OLD_LIFERAY_HOME=$(dirname "$OLD_CATALINA_HOME")
 
-		echo "Unzipping $UNZIP_NAME"
+		if [ "." != "$OLD_LIFERAY_HOME" ]; then
+			for file in $(find $OLD_LIFERAY_HOME -mindepth 1 -maxdepth 1); do
+				mv $file .
+			done
 
-		for file in $(find $UNZIP_NAME -mindepth 1 -maxdepth 1); do
-			mv $file .
-		done
-
-		rmdir $UNZIP_NAME
+			rmdir $OLD_LIFERAY_HOME
+		fi
 	fi
 
-	rm $BUILD_NAME
+	if [ "" != "$BUILD_NAME" ]; then
+		rm $BUILD_NAME
+	fi
 
 	cd -
 }
@@ -173,7 +231,6 @@ makesymlink() {
 	fi
 
 	CATALINA_HOME=$(find ${LIFERAY_HOME} -mindepth 1 -maxdepth 1 -name 'tomcat*')
-	ls -1 ${LIFERAY_HOME}
 	echo "Adding symbolic link to $CATALINA_HOME"
 	ln -s $CATALINA_HOME ${LIFERAY_HOME}/tomcat
 }
@@ -186,6 +243,7 @@ fi
 
 downloadbuild
 makesymlink
+copyextras
 
 # Copy portal-ext.properties, if present
